+++ src/directives/when.ts
@@ -34,12 +34,14 @@ const partCaches = new WeakMap<NodePart, PartCache>();
  *
  * Example:
  *
+ * ```
  * let checked = false;
  *
  * html`
  *   when(checked, () => html`Checkmark is checked`, () => html`Checkmark is not
  * checked`);
  * `
+ * ```
  *
  * @param condition the condition to test truthiness against
  * @param trueValue the value to render given a true condition
@@ -56,8 +58,8 @@ export const when =
             // docment fragment which we cache the nodes of the condition that's
             // not currently rendered.
             cache = {
-              truePart: new NodePart(parentPart.templateFactory),
-              falsePart: new NodePart(parentPart.templateFactory),
+              truePart: new NodePart(parentPart.options),
+              falsePart: new NodePart(parentPart.options),
               cacheContainer: document.createDocumentFragment(),
             };
             partCaches.set(parentPart, cache);