+++ src/lib/template-processor.ts
@@ -14,7 +14,7 @@
 
 import {Part} from './part.js';
 import {NodePart} from './parts.js';
-import {TemplateFactory} from './template-factory.js';
+import {RenderOptions} from './render-options.js';
 
 export interface TemplateProcessor {
   /**
@@ -26,12 +26,13 @@ export interface TemplateProcessor {
    * @param strings The string literals. There are always at least two strings,
    *   event for fully-controlled bindings with a single expression.
    */
-  handleAttributeExpressions(element: Element, name: string, strings: string[]):
-      Part[];
+  handleAttributeExpressions(
+      element: Element, name: string, strings: string[],
+      options: RenderOptions): Part[];
 
   /**
    * Create parts for a text-position binding.
-   * @param templateFactory
+   * @param partOptions
    */
-  handleTextExpression(templateFactory: TemplateFactory): NodePart;
+  handleTextExpression(options: RenderOptions): NodePart;
 }