+++ CHANGELOG.md
@@ -12,9 +12,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- ## Unreleased -->
 <!-- ### Added -->
+## Unreleased
 ### Changed
 * Re-implemented repeat directive for better performance ([#501](https://github.com/Polymer/lit-html/pull/501))
 * Updated TypeScript dependency to 3.1
+* [Breaking] `render()` now takes an options object as the third argument. ([#523](https://github.com/Polymer/lit-html/pull/523))
+### Added
+* Event listeners are called with a configurable `this` reference, which is set via the `eventContext` option to `render()`. ([#523](https://github.com/Polymer/lit-html/pull/523))
+* Support for event listener options, by passing the listener itself as both the second and third arguments to add/removeEventListener().
+
 <!-- ### Removed -->
 <!-- ### Fixed -->